requirements.txt
google-adk>=0.5.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
